+++ tests/stack/test_single_node.py
@@ -7,6 +7,7 @@ from netqasm.lang.instr.flavour import NVFlavour
 from netqasm.lang.parsing import parse_text_subroutine
 from netsquid.components import QuantumProcessor
 from netsquid.qubits import ketstates, qubitapi
+from netsquid_netbuilder.modules.qdevices.generic import GenericQDeviceConfig
 from netsquid_netbuilder.modules.qdevices.nv import NVQDeviceConfig
 from netsquid_netbuilder.util.network_generation import create_single_node_network
 
@@ -16,7 +17,7 @@ from squidasm.sim.stack.common import AppMemory
 from squidasm.sim.stack.host import Host
 
 
-class TestSingleNode(unittest.TestCase):
+class TestSingleNodeNV(unittest.TestCase):
     def setUp(self) -> None:
         ns.sim_reset()
         config = NVQDeviceConfig.perfect_config()
@@ -153,6 +154,58 @@ class TestSingleNode(unittest.TestCase):
 
         self._host = TestHost
         self._check_qmem = check_qmem
+
+
+class TestSingleNodeGeneric(unittest.TestCase):
+    def setUp(self) -> None:
+        ns.sim_reset()
+        config = GenericQDeviceConfig.perfect_config()
+        network_cfg = create_single_node_network(
+            qdevice_typ="generic", qdevice_cfg=config
+        )
+        self.network = _setup_network(network_cfg)
+        self._node = self.network.stacks["Alice"]
+
+        self._host: Optional[Type[Host]] = None
+
+    def tearDown(self) -> None:
+        self._node.subprotocols[f"{self._node.name}_host_protocol"] = self._host(
+            self._node.host_comp
+        )
+        _run(self.network)
+
+    def test_quantum_instructions(self):
+        SUBRT_1 = """
+        # NETQASM 1.0
+        # APPID 0
+        set Q0 0
+        qalloc Q0
+        init Q0
+        x Q0
+        y Q0
+        z Q0
+        h Q0
+        s Q0
+        k Q0
+        t Q0
+        rot_x Q0 16 4
+        rot_y Q0 8 1
+        rot_z Q0 16 2
+        """
+
+        class TestHost(Host):
+            def run(self) -> Generator[EventExpression, None, None]:
+                self.send_qnos_msg(bytes(InitNewAppMessage(max_qubits=2)))
+                app_id = yield from self.receive_qnos_msg()
+                assert app_id == 0
+                subroutine = parse_text_subroutine(SUBRT_1)
+                subroutine.app_id = app_id
+                self.send_qnos_msg(bytes(SubroutineMessage(subroutine)))
+                app_mem = yield from self.receive_qnos_msg()
+                assert isinstance(app_mem, AppMemory)
+                assert app_mem.get_reg_value("Q0") == 0
+
+        self._host = TestHost
 
 
 if __name__ == "__main__":